Fixed some compilation problems with lhs.  Factored out version logic to version.h to reduce dependency on lhs
git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@1369 88ce85bb-8564-4ae6-884f-ab7503230054

+++ simulation/grmain.cpp
@@ -588,7 +588,7 @@ int main(int argc, char** argv)
     return 0;
   }
 
-  std::vector<int> ids;
+  std::vector<size_t> ids;
   {
     std::stringstream ss(track_ids);
     string id;
